% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancytree.R
\name{fancytree}
\alias{fancytree}
\title{fancytree}
\usage{
fancytree(
  id,
  source = NULL,
  ...,
  theme = "win8",
  add.header = TRUE,
  extensions = "",
  auto_unbox = TRUE
)
}
\description{
fancytree
}
