% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancytree.R
\name{fancytreeHeader}
\alias{fancytreeHeader}
\title{Header for fancytree}
\usage{
fancytreeHeader(
  ...,
  extensions = NULL,
  theme = "win8",
  add.filetree.js = TRUE,
  add.context.menu = FALSE
)
}
\description{
Header for fancytree
}
