% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.R
\name{freezeHeaderPage}
\alias{freezeHeaderPage}
\title{A page width a header frozen at the top and scrollbar for the content}
\usage{
freezeHeaderPage(header, ...,
  header.style = "max-height: 5em; overflow: auto; margin-left: 5\%; margin-right: 5\%; padding-bottom: 5px",
  content.style = "overflow: auto;", content.offset = 5)
}
\description{
A page width a header frozen at the top and scrollbar for the content
}

