% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{frozenHeaderPane}
\alias{frozenHeaderPane}
\title{}
\usage{
frozenHeaderPane(
  id = NULL,
  header = NULL,
  content = NULL,
  parent = "body",
  parent.layout = NULL,
  parent.pane = "center",
  style = NULL,
  json.opts = NULL,
  auto.resize = TRUE
)
}
\description{

}
