% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeable.r
\name{hlrPage}
\alias{hlrPage}
\title{A page width a header frozen at the top and scrollbar for the content}
\usage{
hlrPage(header, left, right,
  header.style = "max-height: 5em; overflow: auto; margin-left: 5\%; margin-right: 5\%; padding-bottom: 5px",
  body.style = "overflow: auto;", content.offset = 5,
  freeze.header = TRUE, header.id = "pageHeaderDiv",
  body.id = "pageBodyDiv")
}
\description{
A page width a header frozen at the top and scrollbar for the content
}

