% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{html.image.animation.player}
\alias{html.image.animation.player}
\title{A simple image player for animations}
\usage{
html.image.animation.player(
  container.id,
  base.name = NULL,
  image.files = NULL,
  extension = "svg",
  images.dir = file.path(getwd(), "images"),
  loop.mode = c("none", "loop", "sweep")[1],
  delay = 250,
  add.container = TRUE
)
}
\description{
you can create svg images from multiple plots with record.svg.plots
}
