% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{jqueryLayoutPage}
\alias{jqueryLayoutPage}
\title{A jqueryLayoutPage}
\usage{
jqueryLayoutPage(..., id = "mainPane", center = NULL, north = NULL,
  west = NULL, east = NULL, south = NULL, panes = list(center = center,
  north = north, west = west, east = east, south = south), json.opts = "",
  style = NULL)
}
\description{
A jqueryLayoutPage
}

