% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{jqueryLayoutPanes}
\alias{jqueryLayoutPanes}
\title{A jqueryLayoutPage}
\usage{
jqueryLayoutPanes(center = NULL, north = NULL, west = NULL, east = NULL,
  south = NULL, panes = list(center = center, north = north, west = west,
  east = east, south = south), json.opts = "", style = NULL,
  js.do.layout = TRUE, parent = "body")
}
\description{
A jqueryLayoutPage
}

