% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedSelector.R
\name{nestedSelector}
\alias{nestedSelector}
\title{Nested select menus that show associated div elements}
\usage{
nestedSelector(id, selectors, label = "", show.first = TRUE,
  input.type = c("radioBtnGroup", "select")[1], btn.size = "sm",
  selector.par = list(), scroll.top.sel = NULL)
}
\description{
Nested select menus that show associated div elements
}

