% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedSelector.R
\name{nestedSelectorHandler}
\alias{nestedSelectorHandler}
\title{Add a change handler to a nested selector}
\usage{
nestedSelectorHandler(id, fun, ..., app = getApp())
}
\value{
value a list with the values of all shown selector parts
}
\description{
Add a change handler to a nested selector
}
