% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttongroup.R
\name{radioBtnGroup}
\alias{radioBtnGroup}
\title{A button group that acts like a radio button group}
\usage{
radioBtnGroup(id, labels, values = seq_along(labels), handler = NULL,
  div.style = "", div.extra.class = "", btn.size = "xs", btn.style = "",
  show.hide.containers = NULL, ...)
}
\description{
use radioBtnGroupHandler to add a handler for a value change
}

