% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttongroup.R
\name{radioBtnGroupHandler}
\alias{radioBtnGroupHandler}
\title{Add change hangdler to a radio button group}
\usage{
radioBtnGroupHandler(
  id,
  fun,
  ...,
  eventId = "radioBtnGroupChange",
  app = getApp()
)
}
\value{
value the value of the selected button
}
\description{
Add change hangdler to a radio button group
}
