% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.r
\name{resizeLayout}
\alias{resizeLayout}
\title{Resize a jqueryLayout. May be useful if inner html has been dynamically changed and we have size="auto"}
\usage{
resizeLayout(id, layout.var = paste0(id, "LayoutVar"))
}
\description{
Resize a jqueryLayout. May be useful if inner html has been dynamically changed and we have size="auto"
}
