% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttongroup.R
\name{updateRadioBtnGroup}
\alias{updateRadioBtnGroup}
\title{Change selected button of a radion button group}
\usage{
updateRadioBtnGroup(session = app$session, id, value, app = getApp())
}
\description{
use radioBtnGroupHandler to add a handler for a value change
}

