% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2ui.R
\name{w2sidebar}
\alias{w2sidebar}
\title{Sidebar. See w2ui docs}
\usage{
w2sidebar(
  id,
  img = NULL,
  nodes = NULL,
  ...,
  width = "100\%",
  height = "100\%",
  js.on.render = "",
  add.header = TRUE
)
}
\description{
Sidebar. See w2ui docs
}
