% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2ui.R
\name{w2tabs}
\alias{w2tabs}
\title{Tabs. See w2ui docs}
\usage{
w2tabs(id, active = 1, tabs = NULL, js.on.render = NULL,
  add.header = TRUE)
}
\arguments{
\item{id}{the id of the whole tabset}

\item{active}{the index of the initially shown tab}

\item{tabs}{a list of tabs each tab is itself a list. Example:
tabs=list(list(id="newTab",caption="New Tab", div_id="tab4div", closable=TRUE, keep_closed_content=FALSE)))
\itemize{
\item id - the id of the tab
\item caption - the caption shown on the tab
\item div_id - the id of the div that contains the tab content
\item closable - if TRUE the tab can be closed by pressing on an "x" icon
\item keep_closed_content if FALSE (default) the tab div with div_id will be removed when the tab is closed. If TRUE the div will only be hidden.
}}
}
\description{
Tabs. See w2ui docs
}

